personal: get fully functional and translated
