+++ src/components/ui/badge/index.ts
@@ -1,6 +1,7 @@
 import { cva } from 'class-variance-authority';
 
 export { default as Badge } from './Badge.vue';
+export { default as StickableStatusBadge } from './StickableStatusBadge.vue';
 
 export const badgeVariants = cva(
   'inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2',
