+++ src/i18n/tricks/multiTrickSelect/index.ts
@@ -1,0 +1,5 @@
+import en from './multiTrickSelect.en.json';
+import es from './multiTrickSelect.es.json';
+import fr from './multiTrickSelect.fr.json';
+
+export default { en, es, fr };
