+++ src/i18n/tricks/multiTrickSelect/multiTrickSelect.en.json
@@ -1,0 +1,4 @@
+{
+  "searchPlaceholder": "Add trick...",
+  "noTrickMatchingSearch": "Couldn't find a trick with that name."
+}
