+++ src/i18n/tricks/multiTrickSelect/multiTrickSelect.es.json
@@ -1,0 +1,4 @@
+{
+  "searchPlaceholder": "Añadir truco...",
+  "noTrickMatchingSearch": "No pude encontrar un truco con ese nombre."
+}
