+++ src/i18n/tricks/multiTrickSelect/multiTrickSelect.fr.json
@@ -1,0 +1,4 @@
+{
+  "searchPlaceholder": "Ajouter une astuce...",
+  "noTrickMatchingSearch": "Je n'ai pas trouvé de truc avec ce nom."
+}
