+++ src/i18n/tricks/new/tricksNew.en.json
@@ -35,8 +35,7 @@
     "establishedBy": "Ian, probably",
     "description": "A fitting description",
     "tips": "This is my first tip.\nAnd this is my second tip.",
-    "difficulty": "I'm not sure",
-    "variantOf": "Add trick..."
+    "difficulty": "I'm not sure"
   },
 
   "componentNotImplemented": "This component has yet to be implemented.",
@@ -50,10 +49,6 @@
   "error": {
     "title": "Uh oh…",
     "message": "An error happened when trying to insert the new trick into the DB. Check the console (F12) for more info."
-  },
-
-  "other": {
-    "multiTrickSelectNoTrickFound": "Couldn't find a trick with that name."
   },
 
   "buttonSubmit": "Create",
